WIP try to handle errors
Intermittent failure: got a VCR "no casssette in use" error when running
whole suite for the remove admin status spec (~ line 150)

The issue is maybe that the API reqest happens in a Turbo request. The
idea here is that by putting the VCR block around the whole example, it
covers the turbo update as well.