Don't let occasional organisers set last date
I don't think there's any point? This date never gets used for
occasional events? It's legacy from when we used to check for whether an
event was "active" or not.

I think it's just confusing - people seem to be setting this without
understanding what it means: we're seeing that via errors.

Admins can still set the date, so let's allow organisers to unset it (or
modify it) if it's there (actually removing it is a bigger task), but
otherwise don't show it.