Fix: don't show frequency for weekly class on error
1. Start a new event
2. Select Weekly class
3. click save (get a validation error)

At this point we were in a state we hadn't accounted for: the "When?"
section is shown, but not as a direct result of clicking an event type.

As a result we were showing both frequency options, allowing the user to
select "occasional" and result in an error.

We fix this by applying a hidden class to the frequency section based on
event type.