Venues have static maps
Sometimes people get coordinates very wrong and we end up with a map
marker in Japan (for example).

A simple way to help avoid that is to add a map preview with a static
map.

This stretches the layout a bit on mobile, but these pages aren't really
good on mobile anyway.