Set a max radius for Venues
We sometimes make mistakes and end up with venues with coordinates a
long way away. This sets a radius, which is a blunt instrument, but
should help protect against mistakes.

It's a bit messy to maintain the coords of the Savoy in the smoke spec
(which were wrong btw) but I kind of think it's cute.