WIP hide cancellations on create
https://trello.com/c/dSivRwFg/178-when-creating-an-event-dont-show-a-cancelled-dates-field-for-monthly-events

For monthly/occasional events, the point of “Cancelled dates” is to mark a date as cancelled which has already been listed.

So it doesn’t make any sense to show it when creating an event, since no dates have been listed yet: it’s just another field that might be filled in by accident.

PROBLEM:

We want toggling weekly/occasional to show/hide the field on _create_
but on _update_ we want it to always be visible. This is the only place
in the form where the JS needs to know what mode it's in?