Guard for last date
The Last date field seems to confuse people: it gets filled in when the
event isn't actually ending.

Putting a step in-between, where the user explicitly decides that the
event should be marked as ending, should hopefully help with this a bit.

These changes trip the form objects over the class length limit, but
part of that is due to the to_h hash which I don't think should be made
smaller so I thihnk it's reasonable to add a general CountAsOne modifier
to the cop.

Note that this makes the form longer so we need larger screenshots.