Geocoding uses address
...not just postcode

Unfortunately including name doesn't seem to work unless the name is an
exact match: fortunately this is highlighted by our specific example:
the google maps place entry is named "100 club":

  https://maps.app.goo.gl/f7D2ersihGF8pnUC9

...which is maybe technically correct, but the website is
https://www.the100club.co.uk/, and "The" is used interchangably.

Either way, "The 100 club" can't be geocoded, presumably for this
reason.

...but at least the address should help give more accurate results, and
is hopefully more flexible to variations of rendering the address?