Refactor EventUpdater
This is less idiomatic but more explict: the idiomatic approach is
optimised around nested attributes etc. and rails deletes creates and
updates records in a way which is a bit hard to wrap my head around.

Instead let's explicitly do each step:

1. delete instances which no longer correspond to any dates
2. add or update existing instances
3. update the record (this needs to be last because we have some
   cross-model validations eg. ensuring that all instances are cancelled
   for weekly events)

Wrap it all in a transaction to make sure that the data is saved
consisitently across the two models.