Don't assert `update!` in spec
I want to stop using update! so I need to assert the actual side
effect. In reality any spec failure won't be that the event isn't weekly
- it's that frequency is nil, not 0 or 1.

Not sure how to feel about this, but there are many ways to update a
record, and update! is just one of them, so I don't think it's
appropriate to stub.