WIP Editors can create event hiatuses
TODO: either disallow on create, or implement for create (and add specs).

We're not actually doing anything with them yet, but this drives out the
UI. I could add some assertions that the records are correctly created
in the DB at this point, but we're about to use them and be able to
observe the effects of that, so it doesn't seem worth it.